just playing
roles and privileges

+++ src/main/java/org/teomant/appointment/vote/domain/service/imp/VoteServiceImpl.java
@@ -2,6 +2,9 @@ package org.teomant.appointment.vote.domain.service.imp;
 
 import lombok.RequiredArgsConstructor;
 import org.springframework.stereotype.Service;
+import org.teomant.appointment.security.domain.model.ActionNameEnum;
+import org.teomant.appointment.security.domain.model.EntityNameEnum;
+import org.teomant.appointment.security.service.RightChecker;
 import org.teomant.appointment.vote.domain.model.Vote;
 import org.teomant.appointment.vote.domain.repository.VoteRepository;
 import org.teomant.appointment.vote.domain.service.VoteService;
@@ -11,6 +14,7 @@ import org.teomant.appointment.vote.domain.service.VoteService;
 public class VoteServiceImpl implements VoteService {
 
     private final VoteRepository voteRepository;
+    private final RightChecker rightChecker;
 
     @Override
     public Vote create(Vote vote) {
@@ -19,6 +23,12 @@ public class VoteServiceImpl implements VoteService {
 
     @Override
     public void delete(Vote vote) {
+        Vote stored = voteRepository.findById(vote.getId());
+
+        if (!rightChecker.checkCanPerform(EntityNameEnum.VOTE, ActionNameEnum.DELETE, stored.getUser())) {
+            throw new IllegalArgumentException();
+        }
+
         voteRepository.delete(vote);
     }
 }
